% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_division_peak_graphs.R
\name{generate_division_peak_graphs}
\alias{generate_division_peak_graphs}
\title{Generate Division Peaks Recap Graphs}
\usage{
generate_division_peak_graphs(
  sce,
  moi = "CFSE",
  soi = c("CD8_CM", "CD8_EM", "Th1", "eTreg"),
  conditions = c("stim_0x_Treg", "stim_10x_eTreg", "stim_10x_nTreg", "stim_10x_Tfr",
    "unstim_0x_Treg"),
  patient_id_col = "patient_id",
  subsets_col = "subsets",
  condition_col = "condition"
)
}
\arguments{
\item{sce}{A SingleCellExperiment object containing the data.}

\item{moi}{Marker of interest.}

\item{soi}{Vector of subsets of interest names to filter data.}

\item{conditions}{Vector of condition names to filter data.}

\item{patient_id_col}{Name of the column containing patient IDs.}

\item{subsets_col}{Name of the column containing subset information.}

\item{condition_col}{Name of the column containing condition information.}
}
\value{
A list of graphs recapitulating division peaks per donor.
}
\description{
This function generates graphs that recapitulate division peaks per donor.
}
