% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDivPro.R
\name{scDivPro}
\alias{scDivPro}
\title{Single cell division profiling function}
\usage{
scDivPro(
  sce,
  donor_list,
  division_peaks_graphs_list,
  bd,
  dn,
  num_peaks = 5,
  peak_boundary_uncertainty = 0.2,
  approximate_peak_distance = 1,
  condition = "stim_0x_Treg",
  moi = "CFSE",
  moi2 = "Ki67",
  soi_list = c("CD8_CM", "CD8_EM", "Th1", "eTreg")
)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{donor_list}{Vector containing donor identifiers}

\item{division_peaks_graphs_list}{List of division peaks graphs}

\item{bd}{Beginning dip, as in the dip between division 0 and division 1}

\item{dn}{Donor number}

\item{num_peaks}{Number of peaks.}

\item{peak_boundary_uncertainty}{The uncertainty for peak boundary approximation (default is 0.2)}

\item{approximate_peak_distance}{The approximate distance between peaks (default is 1)}

\item{condition}{Condition of interest.}

\item{moi}{Marker of interest (default is CFSE).}

\item{moi2}{2nd marker of interest to use for scatter plot (default is Ki67).}

\item{soi_list}{A vector containing all subsets of interest.}
}
\value{
A list containing SCE data and graphs related to CFSE analysis.
}
\description{
This function annotates each cell subset by division number and generates accompanying graphs.
}
